requirements.txt
mysql-connector-python==9.0.0
fastapi==0.115.5
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.17